LRAC-10636 move to analytics-dxp-entity-rest-api (buildREST)
