+++ modules/apps/analytics/analytics-dxp-entity-api/build.gradle
@@ -1,14 +1,7 @@
 dependencies {
-	compileOnly group: "com.fasterxml.jackson.core", name: "jackson-annotations", version: "2.12.4"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "io.github.graphql-java", name: "graphql-java-annotations", version: "5.4"
-	compileOnly group: "io.swagger.core.v3", name: "swagger-annotations", version: "2.0.5"
-	compileOnly group: "javax.validation", name: "validation-api", version: "2.0.1.Final"
-	compileOnly group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.1"
-	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
 	compileOnly group: "org.osgi", name: "org.osgi.annotation.versioning", version: "1.1.0"
-	compileOnly project(":apps:portal-odata:portal-odata-api")
+	compileOnly project(":apps:analytics:analytics-dxp-entity-rest-api")
 	compileOnly project(":apps:portal-vulcan:portal-vulcan-api")
 	compileOnly project(":core:petra:petra-function")
-	compileOnly project(":core:petra:petra-string")
 }
