+++ modules/apps/analytics/analytics-dxp-entity-rest-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Analytics DXP Entity REST API
+Bundle-SymbolicName: com.liferay.analytics.dxp.entity.rest.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.analytics.dxp.entity.rest.dto.v1_0
