+++ modules/apps/analytics/analytics-dxp-entity-rest-api/src/main/java/com/liferay/analytics/dxp/entity/rest/dto/v1_0/ExpandoField.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.analytics.dxp.entity.dto.v1_0;
+package com.liferay.analytics.dxp.entity.rest.dto.v1_0;
 
 import com.fasterxml.jackson.annotation.JsonFilter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -20,7 +20,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import com.liferay.petra.function.UnsafeSupplier;
 import com.liferay.petra.string.StringBundler;
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.vulcan.graphql.annotation.GraphQLField;
 import com.liferay.portal.vulcan.graphql.annotation.GraphQLName;
@@ -37,8 +36,6 @@ import java.util.Set;
 
 import javax.annotation.Generated;
 
-import javax.validation.Valid;
-
 import javax.xml.bind.annotation.XmlRootElement;
 
 /**
@@ -46,18 +43,46 @@ import javax.xml.bind.annotation.XmlRootElement;
  * @generated
  */
 @Generated("")
-@GraphQLName("Field")
+@GraphQLName("ExpandoField")
 @JsonFilter("Liferay.Vulcan")
-@XmlRootElement(name = "Field")
-public class Field implements Serializable {
+@XmlRootElement(name = "ExpandoField")
+public class ExpandoField implements Serializable {
 
-	public static Field toDTO(String json) {
-		return ObjectMapperUtil.readValue(Field.class, json);
+	public static ExpandoField toDTO(String json) {
+		return ObjectMapperUtil.readValue(ExpandoField.class, json);
 	}
 
-	public static Field unsafeToDTO(String json) {
-		return ObjectMapperUtil.unsafeReadValue(Field.class, json);
+	public static ExpandoField unsafeToDTO(String json) {
+		return ObjectMapperUtil.unsafeReadValue(ExpandoField.class, json);
 	}
+
+	@Schema
+	public Long getColumnId() {
+		return columnId;
+	}
+
+	public void setColumnId(Long columnId) {
+		this.columnId = columnId;
+	}
+
+	@JsonIgnore
+	public void setColumnId(
+		UnsafeSupplier<Long, Exception> columnIdUnsafeSupplier) {
+
+		try {
+			columnId = columnIdUnsafeSupplier.get();
+		}
+		catch (RuntimeException re) {
+			throw re;
+		}
+		catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	@GraphQLField
+	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
+	protected Long columnId;
 
 	@Schema
 	public String getName() {
@@ -86,18 +111,17 @@ public class Field implements Serializable {
 	protected String name;
 
 	@Schema
-	@Valid
-	public Object getValue() {
+	public String getValue() {
 		return value;
 	}
 
-	public void setValue(Object value) {
+	public void setValue(String value) {
 		this.value = value;
 	}
 
 	@JsonIgnore
 	public void setValue(
-		UnsafeSupplier<Object, Exception> valueUnsafeSupplier) {
+		UnsafeSupplier<String, Exception> valueUnsafeSupplier) {
 
 		try {
 			value = valueUnsafeSupplier.get();
@@ -112,7 +136,7 @@ public class Field implements Serializable {
 
 	@GraphQLField
 	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Object value;
+	protected String value;
 
 	@Override
 	public boolean equals(Object object) {
@@ -120,13 +144,13 @@ public class Field implements Serializable {
 			return true;
 		}
 
-		if (!(object instanceof Field)) {
+		if (!(object instanceof ExpandoField)) {
 			return false;
 		}
 
-		Field field = (Field)object;
+		ExpandoField expandoField = (ExpandoField)object;
 
-		return Objects.equals(toString(), field.toString());
+		return Objects.equals(toString(), expandoField.toString());
 	}
 
 	@Override
@@ -140,6 +164,16 @@ public class Field implements Serializable {
 		StringBundler sb = new StringBundler();
 
 		sb.append("{");
+
+		if (columnId != null) {
+			if (sb.length() > 1) {
+				sb.append(", ");
+			}
+
+			sb.append("\"columnId\": ");
+
+			sb.append(columnId);
+		}
 
 		if (name != null) {
 			if (sb.length() > 1) {
@@ -162,17 +196,11 @@ public class Field implements Serializable {
 
 			sb.append("\"value\": ");
 
-			if (value instanceof Map) {
-				sb.append(JSONFactoryUtil.createJSONObject((Map<?, ?>)value));
-			}
-			else if (value instanceof String) {
-				sb.append("\"");
-				sb.append(_escape((String)value));
-				sb.append("\"");
-			}
-			else {
-				sb.append(value);
-			}
+			sb.append("\"");
+
+			sb.append(_escape(value));
+
+			sb.append("\"");
 		}
 
 		sb.append("}");
@@ -182,7 +210,7 @@ public class Field implements Serializable {
 
 	@Schema(
 		accessMode = Schema.AccessMode.READ_ONLY,
-		defaultValue = "com.liferay.analytics.dxp.entity.dto.v1_0.Field",
+		defaultValue = "com.liferay.analytics.dxp.entity.rest.dto.v1_0.ExpandoField",
 		name = "x-class-name"
 	)
 	public String xClassName;
