+++ modules/apps/analytics/analytics-dxp-entity-rest-api/src/main/resources/com/liferay/analytics/dxp/entity/rest/dto/v1_0/packageinfo
