+++ modules/apps/analytics/analytics-dxp-entity-rest-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/openapi.properties
